Fix login button being cut-off on keyboard shown
As a solution the 'Business in the cloud' message is
hidden when the keyboard is shown.

Also, that message's size is now in dps, so it doesn't get
even bigger when the user uses a larger font size.

Note that Android does not have a way of listening to
the fact whether the keyboard is shown or not, so the library
which is used might not always work. It should work on most
'regular' devices.

Also note that the library is not maintained, although PR's
are still being accepted according to the author.